Bluetooth: Fixed klocwork issues
Fixed klocwork issues
- Null pointer dereference
- Resource Leaks

(cherry picked from commit fffafb6c14fb1298cb2955168a0af893ae7a4fcf)

+++ src/com/android/bluetooth/ftp/FileUtils.java
@@ -71,6 +71,10 @@ public class FileUtils {
         if (D) Log.d(TAG, "deleteDirectory() +");
         if(dir.exists()) {
             File [] files = dir.listFiles();
+            if (files == null) {
+                Log.e(TAG, "error in listing directory ");
+                return false;
+            }
             for(int i = 0; i < files.length;i++) {
                 if(files[i].isDirectory()) {
                     deleteDirectory(mCallback,files[i]);
@@ -105,6 +109,10 @@ public class FileUtils {
          int ret = 0;
          dest.mkdir();
          File [] files = src.listFiles();
+         if (files == null) {
+             Log.e(TAG, "error in listing directory");
+             return ResponseCodes.OBEX_HTTP_INTERNAL_ERROR;
+         }
          for(int i = 0; i < files.length; i++) {
              if (D) Log.d(TAG,"Files =" + files[i]);
              if(files[i].isDirectory()) {
@@ -146,6 +154,15 @@ public class FileUtils {
         } catch(IOException e) {
             Log.e(TAG,"copyFile open stream failed "+ e.toString());
             return ResponseCodes.OBEX_HTTP_INTERNAL_ERROR;
+        } finally {
+            if (null != reader && null == writer) {
+                try {
+                    reader.close();
+                } catch (IOException e) {
+                    Log.e(TAG, "copyFile close stream failed"+ e.toString());
+                    return ResponseCodes.OBEX_HTTP_INTERNAL_ERROR;
+                }
+            }
         }
 
         BufferedInputStream ins = new BufferedInputStream(reader, 0x40000);
